cleanup: remove temporary helper scripts
