--- changed.txt
@@ -1,61 +1,0 @@
-README.md
-scripts/update_skill_readmes.py
-skills/blog-cover-image-cli/README.md
-skills/brand-alchemy/README.md
-skills/claude-md-generator/README.md
-skills/cold-email-verifier/README.md
-skills/company-radar/README.md
-skills/competitor-pr-finder/README.md
-skills/cook-the-blog/README.md
-skills/dependency-update-bot/README.md
-skills/docs-from-code/README.md
-skills/domain-expired-opportunity-finder/README.md
-skills/dx-roaster/README.md
-skills/email-newsletter/README.md
-skills/explain-this-pr/README.md
-skills/gh-issue-to-demand-signal/README.md
-skills/github-discussion-to-devrel-content/README.md
-skills/google-trends-api-skills/README.md
-skills/graphic-case-study/README.md
-skills/graphic-chart/README.md
-skills/graphic-ebook/README.md
-skills/graphic-gif/README.md
-skills/graphic-slide-deck/README.md
-skills/hackernews-intel/README.md
-skills/human-tone/README.md
-skills/hyperframes-product-launch-video/README.md
-skills/kill-the-standup/README.md
-skills/linkedin-job-post-to-buyer-pain-map/README.md
-skills/linkedin-post-generator/README.md
-skills/llms-txt-generator/README.md
-skills/map-your-market/README.md
-skills/meeting-brief-generator/README.md
-skills/meta-ads-skill/README.md
-skills/meta-tribeV2-skill/README.md
-skills/newsletter-digest/README.md
-skills/noise-to-linkedin-carousel/README.md
-skills/noise2blog/README.md
-skills/npm-downloads-to-leads/README.md
-skills/oss-launch-kit/README.md
-skills/outreach-sequence-builder/README.md
-skills/position-me/README.md
-skills/pr-description-writer/README.md
-skills/pricing-finder/README.md
-skills/pricing-page-psychology-audit/README.md
-skills/product-update-logger/README.md
-skills/producthunt-launch-kit/README.md
-skills/reddit-icp-monitor/README.md
-skills/reddit-post-engine/README.md
-skills/schema-markup-generator/README.md
-skills/sdk-adoption-tracker/README.md
-skills/show-hn-writer/README.md
-skills/store-listing-optimizer/README.md
-skills/tweet-thread-from-blog/README.md
-skills/twitter-GTM-find-skill/README.md
-skills/vc-curated-match/README.md
-skills/vc-finder/README.md
-skills/vid-motion-graphics/README.md
-skills/vid-product-launch/README.md
-skills/vid-sizzle-reel/README.md
-skills/where-your-customer-lives/README.md
-skills/yc-intent-radar-skill/README.md
