--- run_git.bat
@@ -1,3 +1,0 @@
-@echo off
-cd /d "C:\Users\Fariz\OneDrive\Desktop\experiments\opendirectory"
-C:\Program Files\Git\cmd\git.exe status --short
